synapse-core/packages/marketing-agent/pyproject.toml
[tool.poetry]
name = "marketing-agent"
version = "0.1.0"
description = ""
authors = ["Codex CLI <travisgray@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.128.0,<0.129.0"
pydantic = ">=2.12.5,<3.0.0"
langchain-core = ">=1.2.5,<2.0.0"
uvicorn = ">=0.40.0,<0.41.0"
python-dotenv = ">=1.2.1,<2.0.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
langchain-openai = ">=1.1.6,<2.0.0"
alembic = ">=1.17.2,<2.0.0"
python-json-logger = ">=4.0.0,<5.0.0"
pyjwt = ">=2.10.0,<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-mock = ">=3.15.1,<4.0.0"
black = ">=25.12.0,<26.0.0"
flake8 = ">=7.3.0,<8.0.0"
httpx = ">=0.28.1,<0.29.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
